:see_no_evil:Update .gitignore
Added client_secrets.json in preperation for using google login.
I will use google to verify a user but also store data such as
name email etc in a database.

Small update to index.html that adds a google login button for
testing as well as 3 buttons that link to the different pages.