:rocket:First FastAPI endpoint succesfully ran
This endpoint is called when a user logs in with google. It will
be passed a google_id and check the database for it and return
true or false. If true then the user is logged in as normal. If
false then then user is brought to the registration page to get
more details.

Next I want to store these users in the users collection