Navbar and video viewing layout
Removed unused bootstrap links as I have the files locally, linked
to these files instead. The navbar is almost complete I just need
to bring more use to the dropdown menu besides loggin out. I am
thinking an account view link and a admin dashboard link for
admin users. When this page loads I fetch the users data from
their google id in session and display it. I need to change this
to use the mongo ID so that all users work with this feature.