Convert and store default profile picture to base64
When a user registers with google, the image link is stored in
session, if email registration occurs then the profile picture
is marked as default. When adding a user to collection I check
if the passed users profile picture = default and start the
conversion. If a user registers with google then the url is stored
and no conversion takes place