:rocket:Displaying live feed from AWS Kinesis to frontend
I have successfully setup my Pi 5 as a producer for my kinesis
live stream. To access the stream I need an HLS URL which can
be aquired via boto3 which creates an aws kinesis client
I can interact with to request this URL. The URL is passed into
the template each time it is reloaded. Chrome does not support
native HLS so I have a smalls crupt inside the home template to
deal with this. I have not setup automatic refresh so once the
URL's time to live is up the video pauses until another refresh