Created voice recording feature to communicate with oscar
With time constraints I do not think I will get this fully
completed. Currently the code allows the user to record a small
voice recording and it uploads it to an uploads/ folder on the
server. My plan would then be to store this in a collection
and have a script running on the pi witch watches the collection
for new entries. If there was a new entry then the audio file
would play. This feature is available from the mic button and
uses a modal to display