Displaying screenshots on profile page
When a user takes a screenshot they can see it on their profile
page. I do not have the time to add the ability to delete these
screenshots but I intend to add this in the future. The last thing
i want to do is hash the passwords before storing them